Dependabot-ify DependencyGrapher as a new component
